Fixed token modal with new action data
