+++ app/components/Modals/TokenModal/index.js
@@ -1,1 +1,11 @@
-export { default } from './TokenModal'
+// @flow
+import { compose } from 'recompose'
+
+import TokenModal from './TokenModal'
+import withNetworkData from '../../../hocs/withNetworkData'
+import withTokensData from '../../../hocs/withTokensData'
+
+export default compose(
+  withNetworkData(),
+  withTokensData()
+)(TokenModal)
