+++ app/containers/App/Header/NetworkSwitch/NetworkSwitch.jsx
@@ -1,7 +1,8 @@
 // @flow
 import React, { Component } from 'react'
 
-import headerStyles from '../Header.scss'
+import { getNetworks } from '../../../../core/networks'
+import styles from '../Header.scss'
 
 type Props = {
   networkId: string,
@@ -10,11 +11,15 @@ type Props = {
 }
 
 export default class NetworkSwitch extends Component<Props> {
+  static defaultProps = {
+    networks: getNetworks()
+  }
+
   render () {
     const { networkId, networks } = this.props
     return (
-      <div id='network' className={headerStyles.navBarItem}>
-        <span className={headerStyles.navBarItemLabel}>Running on</span>
+      <div id='network' className={styles.navBarItem}>
+        <span className={styles.navBarItemLabel}>Running on</span>
         <select defaultValue={networkId} onChange={this.handleChange} className='networkSelector'>
           {networks.map(({ label, id }: NetworkItemType) =>
             <option key={`networkOption${id}`} value={id}>{label}</option>
