+++ app/hocs/withTokensData.js
@@ -1,17 +1,7 @@
 // @flow
-import { compose } from 'recompose'
-
 import withData from './api/withData'
-import withNetworkData from './withNetworkData'
 import settingsActions from '../actions/settingsActions'
 
 export default function withTokensData (key: string = 'tokens') {
-  const mapSettingsDataToProps = (settings, props: Object) => ({
-    [key]: settings.tokens.filter(({ networkId }) => networkId === props.networkId)
-  })
-
-  return compose(
-    withNetworkData(),
-    withData(settingsActions, mapSettingsDataToProps)
-  )
+  return withData(settingsActions, (settings) => ({ [key]: settings.tokens }))
 }
